Renames RigidBody.h/cpp --> rigid_body.h/cc
