+++ drake/multibody/rigid_body.h
